requirements.txt
fastapi
uvicorn
openai
python-dotenv
vosk
yt-dlp
aiofiles
requests
python-multipart
pillow
youtube-transcript-api
pyaudio
